refactor : object 저장소 생성 후 코드 분리
