+++ src/pages/BuildingElectricity/BuildingElectricity.tsx
@@ -6,40 +6,11 @@ import * as S from './BuildingElectricity.style';
 import { Chart as ChartJS, Tooltip, Legend } from 'chart.js/auto';
 import { Line } from 'react-chartjs-2';
 import Carousel from '../../components/Carousel/Carousel';
-import 본관 from '../../assets/schoolImage/본관.jpg';
-import 주년60 from '../../assets/schoolImage/60주년.jpg';
-import { buildingInfoType } from '../../type/Types';
 import downArrow from '../../assets/svg/downArrow.svg';
 import test from '../../api/test';
+import { monthlyInitData, buildingList, buildingCode } from '../../store/store';
 
 ChartJS.register(Tooltip, Legend);
-
-const buildingCode: any = Object.freeze({ 본관: 1, '60주년': 2 });
-
-const data: any = {
-  labels: [
-    '1월',
-    '2월',
-    '3월',
-    '4월',
-    '5월',
-    '6월',
-    '7월',
-    '8월',
-    '9월',
-    '10월',
-    '11월',
-    '12월',
-  ],
-
-  datasets: [
-    {
-      type: 'bar',
-      backgroundColor: 'rgb(75, 192, 192)',
-      data: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
-    },
-  ],
-};
 
 const options = {
   reponsive: false,
@@ -65,29 +36,23 @@ const options = {
 const Chart = ({ chartState }: { chartState: any }) => {
   return (
     <S.Container>
-      <Line width="350" height="200" data={chartState} options={options}></Line>
+      <Line width="350" height="300" data={chartState} options={options}></Line>
     </S.Container>
   );
 };
-
-const buildingList: buildingInfoType[] = [
-  { buildingName: '60주년', src: 주년60 },
-  { buildingName: '본관', src: 본관 },
-];
 
 const BuildingElectricity = () => {
   const [selectedBuilding, setSelectedBuilding] = useState<string>(
     buildingList[0].buildingName
   );
-  const [chartState, setChartState] = useState(data);
+  const [chartState, setChartState] = useState(monthlyInitData);
 
   const testAPI = async (selectedBuilding: number) => {
     const rData = await test(selectedBuilding);
-    console.log(rData);
+
     // 깊은 복사를 하지 않으면 chartJS서 변동 감지를 못함 JSON.parse, JSON.stringify로 깊은 복사
     const chartStateCopy = JSON.parse(JSON.stringify(chartState));
     chartStateCopy.datasets[0].data = rData.result[0].usages;
-
     setChartState(chartStateCopy);
   };
